requirements.txt
absl-py==0.6.1
certifi==2018.11.29
chardet==3.0.4
idna==2.8
kiwishot==1.0
Pillow==5.4.1
pkg-resources==0.0.0
pyimgur==0.6.0
python-xlib==0.23
requests==2.21.0
six==1.12.0
urllib3==1.24.1
